Fix issue #20
Switch from a stateless react component to a stateful component so this.state and this.setState can be used. The avatar component will now check if it's been passed a image url, if not it will fallback to the default. Also if the image url that was passed is invalid and or fails to load avatar will switch to the default image.